+++ libfreeipmi/src/api/ipmi-kcs-driver-api.h
@@ -28,13 +28,13 @@
    highly recommended to use ipmi_kcs_cmd instead. Otherwise make sure
    you check the return status of write before calling read.
 */
-int8_t ipmi_kcs_cmd_api (ipmi_ctx_t ctx,
-                         fiid_obj_t obj_cmd_rq,
-                         fiid_obj_t obj_cmd_rs);
+int ipmi_kcs_cmd_api (ipmi_ctx_t ctx,
+                      fiid_obj_t obj_cmd_rq,
+                      fiid_obj_t obj_cmd_rs);
 
-int8_t ipmi_kcs_cmd_api_ipmb (ipmi_ctx_t ctx,
-                              fiid_obj_t obj_cmd_rq,
-                              fiid_obj_t obj_cmd_rs);
+int ipmi_kcs_cmd_api_ipmb (ipmi_ctx_t ctx,
+                           fiid_obj_t obj_cmd_rq,
+                           fiid_obj_t obj_cmd_rs);
 
 int32_t ipmi_kcs_cmd_raw_api (ipmi_ctx_t ctx,
                               uint8_t *buf_rq,
