+++ libfreeipmi/src/api/ipmi-lan-interface-api.h
@@ -23,9 +23,9 @@
 #include <freeipmi/api/ipmi-api.h>
 #include <freeipmi/fiid/fiid.h>
 
-int8_t ipmi_lan_cmd (ipmi_ctx_t ctx,
-                     fiid_obj_t obj_cmd_rq,
-                     fiid_obj_t obj_cmd_rs);
+int ipmi_lan_cmd (ipmi_ctx_t ctx,
+                  fiid_obj_t obj_cmd_rq,
+                  fiid_obj_t obj_cmd_rs);
 
 int32_t ipmi_lan_cmd_raw (ipmi_ctx_t ctx,
                           uint8_t *buf_rq,
@@ -33,9 +33,9 @@ int32_t ipmi_lan_cmd_raw (ipmi_ctx_t ctx,
                           uint8_t *buf_rs,
                           size_t buf_rs_len);
 
-int8_t ipmi_lan_2_0_cmd (ipmi_ctx_t ctx,
-                         fiid_obj_t obj_cmd_rq,
-                         fiid_obj_t obj_cmd_rs);
+int ipmi_lan_2_0_cmd (ipmi_ctx_t ctx,
+                      fiid_obj_t obj_cmd_rq,
+                      fiid_obj_t obj_cmd_rs);
 
 int32_t ipmi_lan_2_0_cmd_raw (ipmi_ctx_t ctx,
                               uint8_t *buf_rq,
