+++ libfreeipmi/src/api/ipmi-openipmi-driver-api.h
@@ -23,13 +23,13 @@
 #include <freeipmi/api/ipmi-api.h>
 #include <freeipmi/fiid/fiid.h>
 
-int8_t ipmi_openipmi_cmd_api (ipmi_ctx_t ctx,
-                              fiid_obj_t obj_cmd_rq,
-                              fiid_obj_t obj_cmd_rs);
+int ipmi_openipmi_cmd_api (ipmi_ctx_t ctx,
+                           fiid_obj_t obj_cmd_rq,
+                           fiid_obj_t obj_cmd_rs);
 
-int8_t ipmi_openipmi_cmd_api_ipmb (ipmi_ctx_t ctx,
-                                   fiid_obj_t obj_cmd_rq,
-                                   fiid_obj_t obj_cmd_rs);
+int ipmi_openipmi_cmd_api_ipmb (ipmi_ctx_t ctx,
+                                fiid_obj_t obj_cmd_rq,
+                                fiid_obj_t obj_cmd_rs);
 
 int32_t ipmi_openipmi_cmd_raw_api (ipmi_ctx_t ctx,
                                    uint8_t *buf_rq,
